# banksampah_restful_api-openapi.yaml
openapi: "3.0.3"
info:
  title: "banksampah_restful_api"
  description: "API Spec Untuk Aplikasi Bank Sampah (BS) Mobile"
  version: "1.0.0"
servers:
  - url: "http://127.0.0.1:8080"
    description: "Local server"
paths:
  /api/storing-trash:
    post:
      tags:
        - "Storing Trash"
      summary: "Create Storing Trash"
      operationId: "createStoringTrash"
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                accountId:
                  type: "integer"
                  example: 1
                trashId:
                  type: "integer"
                  example: 1
                locationId:
                  type: "integer"
                  example: 1
                weight:
                  type: "integer"
                  example: 1
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        example: 1
                      accountId:
                        type: "object"
                        example: {}
                      trashId:
                        type: "object"
                        example: {}
                      locationId:
                        type: "object"
                        example: {}
                      weight:
                        type: "integer"
                        example: 1
                      total:
                        type: "integer"
                        example: 1
                      createdAt:
                        type: "string"
                        example: "2023-01-14T14:36:10.000+00:00"
                      status:
                        type: "string"
                        example: "DalamProses"
  /api/storing-trash/account/{accountId}:
    get:
      tags:
        - "Storing Trash"
      summary: "Get List Storing Trash By Account Id"
      operationId: "getStoringTrashByAccountId"
      parameters:
        - name: "accountId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "array"
                    items:
                        type: "object"
                        properties:
                            id:
                              type: "integer"
                              example: 1
                            accountId:
                              type: "object"
                              example: {}
                            trashId:
                              type: "object"
                              example: {}
                            locationId:
                              type: "object"
                              example: {}
                            weight:
                              type: "integer"
                              example: 1
                            total:
                              type: "integer"
                              example: 1
                            createdAt:
                              type: string
                              example: "2023-01-14T14:36:10.000+00:00"
                            status:
                              type: "string"
                              example: "DalamProses"
  /api/storing-trash/{id}:
    get:
      tags:
          - "Storing Trash"
      summary: "Get One Storing Trash By Id"
      operationId: "getStoringTrashById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        example: 1
                      accountId:
                        type: "object"
                        example: {}
                      trashId:
                        type: "object"
                        example: {}
                      locationId:
                        type: "object"
                        example: {}
                      weight:
                        type: "integer"
                        example: 1
                      total:
                        type: "integer"
                        example: 1
                      createdAt:
                        type: string
                        example: "2023-01-14T14:36:10.000+00:00"
                      status:
                          type: "string"
                          example: "DalamProses"

    put:
      tags:
          - "Storing Trash"
      summary: "Update Storing Trash by"
      operationId: "updateStoringTrash"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                weight:
                  type: "integer"
                  example: 2
                status:
                  type: "string"
                  example: "Selesai"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        example: 1
                      accountId:
                        type: "object"
                        example: { }
                      trashId:
                        type: "object"
                        example: { }
                      locationId:
                        type: "object"
                        example: { }
                      weight:
                        type: "integer"
                        example: 2
                      total:
                        type: "integer"
                        example: 2
                      createdAt:
                        type: string
                        example: "2023-01-14T14:36:10.000+00:00"
                      status:
                        type: "string"
                        example: "Selesai"

    delete:
      tags:
          - "Storing Trash"
      summary: "Delete Storing Trash by Id"
      operationId: "deleteStoringTrash"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    format: "int32"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "integer"
                    format: "int32"
                    example: 1
...
  /api/trash:
    post:
      tags:
        - "Trash"
      summary: "Create Trash"
      operationId: "createTrash"
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                name:
                  type: string
                  example: "Trash 1"
                price:
                  type: number
                  example: 10000
                description:
                  type: string
                  example: "Trash 1"
                imageId:
                  type: number
                  example: 1
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    format: "int32"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        format: "int32"
                        example: 1
                      name:
                        type: string
                        example: "Trash 1"
                      price:
                        type: number
                        example: 10000
                      description:
                        type: string
                        example: "Trash 1"
                      imageId:
                        type: number
                        example: 1
    get:
      tags:
        - "Trash"
      summary: "Get List Trash"
      operationId: "listTrash"
      parameters:
        - name: "size"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
            default: "100"
        - name: "page"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
            default: "0"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    format: "int32"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "array"
                    items:
                      type: "object"
                      properties:
                        id:
                          type: "integer"
                          format: "int32"
                          example: 1
                        name:
                          type: string
                          example: "Trash 1"
                        price:
                          type: number
                          example: 10000
                        description:
                          type: string
                          example: "Trash 1"
                        imageId:
                          type: number
                          example: 1
  /api/trash/{id_trash}:
    get:
      tags:
        - "Trash"
      summary: "Get One Trash"
      operationId: "getTrash"
      parameters:
        - name: "id_trash"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    format: "int32"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        format: "int32"
                        example: 1
                      name:
                        type: string
                        example: "Trash 1"
                      price:
                        type: number
                        example: 10000
                      description:
                        type: string
                        example: "Trash 1"
                      imageId:
                        type: number
                        example: 1
    put:
      tags:
        - "Trash"
      summary: "Update Trash"
      operationId: "updateTrash"
      parameters:
        - name: "id_trash"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                name:
                  type: string
                  example: "Trash 1"
                price:
                  type: number
                  example: 10000
                description:
                  type: string
                  example: "Trash 1"
                imageId:
                  type: number
                  example: 1
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    format: "int32"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        format: "int32"
                        example: 1
                      name:
                        type: string
                        example: "Trash 1"
                      price:
                        type: number
                        example: 10000
                      description:
                        type: string
                        example: "Trash 1"
                      imageId:
                        type: number
                        example: 1
    delete:
      tags:
        - "Trash"
      summary: "Delete Trash"
      operationId: "deleteTrash"
      parameters:
        - name: "id_trash"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    format: "int32"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "integer"
                    format: "int32"
                    example: 1